+++ apps/core/admin.py
@@ -1,7 +1,8 @@
 from django.contrib import admin
 
-from apps.core.models import EmailSent, Project, ProjectStateTransition
+from apps.core.models import EmailSent, Project, ProjectStateTransition, ProjectSyncRequest
 
 admin.site.register(EmailSent)
 admin.site.register(Project)
 admin.site.register(ProjectStateTransition)
+admin.site.register(ProjectSyncRequest)